+++ assets/js/theme/common/product-details.js
@@ -92,7 +92,16 @@ export default class ProductDetails extends ProductDetailsBase {
         });
 
         this.updateProductAttributes(productAttributesData);
-        this.updateView(productAttributesData, null);
+
+        if (productAttributesData
+            && typeof productAttributesData === 'object'
+            && Object.keys(productAttributesData).length > 0) {
+            this.updateView(productAttributesData, null);
+        } else {
+            // eslint-disable-next-line no-console
+            console.warn('BCData.product_attributes is empty on product initialization');
+        }
+
         bannerUtils.dispatchProductBannerEvent(productAttributesData);
 
         $productOptionsElement.show();